requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
azure-cognitiveservices-speech==1.31.0
gunicorn==20.1.0
api
requests
elevenlabs
elevenlabs
elevenlabs
httpx
pydantic
pydantic-core
requests
typing_extensions
websockets
python-dotenv
requests
